update-ws updated

+++ update-ws.sh
@@ -13,17 +13,37 @@ vcs pull src
 rosdep install -y --from-paths src --ignore-src --rosdistro $ROS_DISTRO
 
 while true; do
-    read -p "Want to rebuild? (y/n)?" yn
+    read -p "Want to rebuild? (y/n)? " yn
     case $yn in
         [Yy]* ) break;;
         [Nn]* ) exit;;
-        * ) echo "Please answer yes or no.";;
+        * ) echo "Please answer yes (y or Y) or no (n or N).";;
     esac
 done
 
+while true; do
+    read -p "Want to delete build/install/log folders beforehand? (y/n)? " yn
 
-rm -R build
-rm -R install
-rm -R log
+	if [[ $yn =~ ^[Yy]$ ]]
+	then
+		echo "Removing ..."
+	    	rm -R build
+		rm -R install
+		rm -R log
+		echo "Removing successful ..."
+		echo "Building ..."
+		colcon build --symlink-install --cmake-args -DCMAKE_BUILD_TYPE=Release
+		echo "Building successful..."
+		break
+	else
+		echo "Building ..."
+		colcon build --symlink-install --cmake-args -DCMAKE_BUILD_TYPE=Release
+		echo "Building successful..."
+		break
+	fi
+done
 
-colcon build --symlink-install --cmake-args -DCMAKE_BUILD_TYPE=Release
+echo "Script is completed ..."
+
+
+
